+++ tom_d4rt_ast/lib/src/runtime/interpreter_visitor.dart
@@ -807,7 +807,9 @@ class InterpreterVisitor extends GeneralizingSAstVisitor<Object?> {
         }
         // If neither getter, method, nor Object property, error
         throw UndefinedMemberD4rtException(
-          "Undefined property or method '$name' on bridged instance of '${bridgedInstance.bridgedClass.name}' accessed via implicit 'this'.",
+          "Undefined property or method '$name' on bridged instance of "
+          "'${bridgedInstance.bridgedClass.name}' accessed via implicit 'this'."
+          "${unbridgedMemberSuffix(bridgedInstance.bridgedClass.name, name)}",
           memberName: name,
         );
       } // +++ NEW BLOCK +++
@@ -1484,7 +1486,9 @@ class InterpreterVisitor extends GeneralizingSAstVisitor<Object?> {
       }
 
       throw D4rtNoSuchMethodError(
-        "Undefined property or method '$memberName' on bridged instance of '${bridgedInstance.bridgedClass.name}'.",
+        "Undefined property or method '$memberName' on bridged instance of "
+        "'${bridgedInstance.bridgedClass.name}'."
+        "${unbridgedMemberSuffix(bridgedInstance.bridgedClass.name, memberName)}",
       );
     } else if (prefixValue is InterpretedRecord) {
       // Accessing field of a record
@@ -4405,7 +4409,9 @@ class InterpreterVisitor extends GeneralizingSAstVisitor<Object?> {
               return extensionMethod.call(this, extensionArgs, namedArgs);
             } else {
               throw D4rtNoSuchMethodError(
-                "Bridged class '${bridgedClass.name}' has no instance method named '$methodName'.",
+                "Bridged class '${bridgedClass.name}' has no instance "
+                "method named '$methodName'."
+                "${unbridgedMemberSuffix(bridgedClass.name, methodName)}",
               );
             }
           } on RuntimeD4rtException catch (findError) {